Finish the refactoring of the comment javascript into coffescript classes. Minor updates to wording and styles along the way.
The comment javascript is now handled by an instantiation of the CommentFormHandler class. This serves as the director for the other classes under javascripts/components/comments/. I've attempted to break down responsibilities into classes that use the CommentFormHandler class as the gateway for communication to the other classes as needed. A CommentFormDecorator class should probably be broken out of the  CommentFormHandler class (to handle the adding of 'public' icons, etc.)