Move token generation into the attachment uploader and use the token to uniquely name the file. This prevents files with the same name from overwriting each other on s3.
This was leading to errors when the one of the comment_attachments was deleted (removing the only file on s3) but the other wasn’t and then the form was submitted. As there was no longer a file on s3 we’d blow up when attempting to download the file.