Allow rendering of precompiled html from staging.
Our current version of varnish has trouble with https and following redirects when dealing with ESI’s - this patch makes it possible to render the compiled html for the document toc and full text from the staging server (as we’re often using the staging api for data in development). Once we upgrade varnish we should be able to remove this.