Revert "Change session paths references to be nested under /my in JS handlers, emails, sign in helper, and CORS initializers so that subscription, comment, and clippings metadata stored in the session will not be destroyed by Varnish."
Though the oauth callback url is technically customizable in at least some of the oauth gems that we’re using, it may not be customizable with the combination of gems that we’re using (e.g. doorkeeper, doorkeeper open id connect, and omniauth as a dependency).  We were encountering issues related to the auth_hash being sent to sessions#create as nil and issues with the integration testing helper not being configurable.  Thus, in order to preserve the cookies, a Varnish path exemption will be added.

This reverts commit c37fe9985907e27c532a27383ffa6aa70d35df01.