Add a check to the external link checker that allows the current domain. Refactor check syntax to be more succinct.
This means we no longer need to whitelist our own domain and it works 
automatically for both staging and production. We precompile assets 
under production in our containers and so using the settings by 
environment doesn't work as a strategy. Left developement settings in 
place as those aren't precompiled and we want to allow prod and staging 
links in dev.