Update: Strip commas when guessing slugs for unrecognized agencies.
When generating the JSON for the table of contents, on occasion the agency is unrecognized and so we attempt to create a simple ‘generated’ slug based on our best guess of what the actual slug is.

In one example, we were generating a slug based on an agency name that included a comment and generated this incorrect slug: “inspector-general-office,-health-and-human-services-department”.  The actual slug was “inspector-general-office-health-and-human-services-department”.

Thus, going forward we will remove commas when attempting to guess agency names.

Related to ticket #120 in API Core.