Bugfix: Fix beta-CSP reporting subdomain and replace with current equivalent
The k8ds staging environment didn’t seem to be generating CSP errors in Honeybadger.  We were involved in the beta for Honeybadger’s CSP, which seems to have used a different subdomain.  It still seems possible to manually POST a CSP violation payload to the old URI we were using, but when hooked up to Rails, it didn’t seem to be reporting CSP violations.

https://docs.honeybadger.io/guides/errors.html#content-security-policy-reports