Update: 404 instead of 500 on missing blog pages.
We already have a rescue that raises a 404 if an item is not located, but it’s being invoked in the view context instead of the controller—as such, the page is rendering a 500 instead of a 404.  By calling the item in initialize we can raise the 404 in the controller context if appropriate.

eg /reader-aids/office-of-the-federal-register-blog//author/aotovo