fix persisting comments after login
non-logged in users will not have a valid CSRF token when commenting. the default CSRF protection strategy is "null_session", which causes all session updates during the request to be ignored. this means 1. that no cookie will be returned for non-logged in users on the create action, and the comment receipt page's CSRF token will thus be invalid, which means the persist_for_login action won't populate a session cookie with the information needed for after login.

by switching to a reset_session strategy, the old session (which likely doesn't exist) is thrown out during step 1, which makes the subsequent calls succeed as expected.