Update: Improve handling of comment file upload errors
Extract custom add function behavior to a processActions function.
Add a reference to CommentFormFileUploader instance to uploader object for retrieval in processAction.
Prepare AJAX options after 'fileuploadsubmit' is triggered and after all processing and error handling is complete.
Add a new binding for 'fileuploadprocessdone' to update list of uploaded files.