Update: First pass at js fix for big doc page perf
Large document pages are experiencing layout thrashing on page load where the browser is indicating to the user that page load is taking a long time.  The slow performance appears to be in the print_page_elements JS class where hundreds of DOM API reflow calls are being executed iteratively instead of being grouped together to minimize reflows.

Example page:
/documents/2021/08/13/2021-16519/medicare-program-hospital-inpatient-prospective-payment-systems-for-acute-care-hospitals-and-the