Fix: FR Index section collapsing.
The rendering of the documents at various levels in the TOC was not uniformly applying the same sorting algorithm (which is document-dependent) when selecting the FIRST document used for specifying the JS target selector AND the actual order of the documents.

Changed the logic here so we're applying the same sorting algorithm when specifying the order of the documents when rendering the fr index subject header (which is then used to calculate the FIRST document for the 'toggleTarget' attribute) AND when actually iterating through the documents for display.

See commit #aa4c429734 for additional context on why the sorting algorithm was applied to beginw with.

See #333.