Update: Popular document algorithm
Post upgrade to Mysql 5.7, the prior popular document algorithm would be executed several times simultaneously in Mysql (perhaps due to a flood of traffic on the site) and the mysql query times for this method would start taking between 40-60 seconds.

We chose to simplify the algorithm, make it more performant, and fix an isssue since more than 1000 documents are published in the last 30 days.

Delete no-longer-appropricate specs.