Fix: Agency index pages caused by non-determinance
We would regularly receive errors related to the index page being unable to find a document for agencies with more than 1000 documents.

We did not have an explicit sort order specified in this query, and so by default, API Core was sorting first by the publication date and then by the score.

In production, some documents would be missing and those documents tended to be right at the pagination cutoff (eg 999, 2999, etc), which pointed towards there being a slight issue with non-deterministic sorting.

In lieu of relying on a sorting order that has any sort of scoring element mixed in, we're now explicitly specifying an id-based sorting order to remove the scoring element completely from the sort to get reliable result order.