Fix: Include sub-agency ids for doc query
It appears the query for building the FR index JSON includes sub-agencies and the front-end query for retrieving documents only included the parent agency.  Use the agency API to look up the parent agency id and its descendant agency ids and use this in the search in lieu of just searching by the parent agency.