Update: Adapt mobile for full-width explore sections
In desktop width, the explore sections use a hanging indent, and the content below the header of the .with-icon hanging indents is not wrapped in a containing element.  Hence there was a need to use the CSS general sibling combinator (~) in order to effectively select all the elements following the heading and apply negative margin to them and set their widths to be the sum of 100% plus the negative margin added.