Refactor: Move all CSS related to pages in transition to separate file.
The different handling we have in place for facilitating the pages that have yet to be transitioned to mobile was becoming unwieldy since it was found in several different places.  This CSS relates to handling pages in transition that have not yet been made fully mobile-friendly.  Once the entire site has been migrated to mobile, we should remove this file.